backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Environment and config
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# HTTP client
httpx>=0.26.0

# Google OAuth and Gmail
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.116.0

# Google Gemini AI
google-generativeai>=0.4.1

# JWT for sessions
pyjwt>=2.8.0

# Testing
pytest>=8.0.0
pytest-mock>=3.12.0
pytest-asyncio>=0.23.4
